requirements.txt
asgiref==3.2.7
Django==3.0.6
django-extensions==2.2.9
django-redis==4.11.0
djangorestframework==3.11.0
factory-boy==2.12.0
Faker==4.1.0
mysqlclient==1.4.6
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.2
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
Werkzeug==1.0.1
